Add second Cr2DecompressorFuzzer with dummy huffman table
Fuzzer *still* seems to having a real bad time reaching the
subsampled variation, let's help it even more by removing
yet another part of the equasion.
