+++ fuzz/all-fuzzers.txt
@@ -1,6 +1,7 @@
 CiffParserFuzzer-GetDecoder
 CiffParserFuzzer-GetDecoder-Decode
-Cr2DecompressorFuzzer
+Cr2DecompressorFuzzer-DummyHuffmanTable
+Cr2DecompressorFuzzer-HuffmanTable
 Cr2LJpegDecoderFuzzer
 CrwDecompressorFuzzer
 DngOpcodesFuzzer
